new view for files
View that links both table od_file and od_data_media to a specific object. To be used in QGIS to visualize files such as photos, pdf aso. There should be some modification on the demoproject in order to use the view instead of the File formular.

This sql needs improvement to add the triggers to insert/delete files.